Move progress bar in components folder
