+++ front/app/components/InitiativeCard/ReactionIndicator.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled, { useTheme } from 'styled-components';
 
 import { Icon, fontSizes, colors } from '@citizenlab/cl2-component-library';
-import ProposalProgressBar from 'containers/InitiativesShow/ReactionControl/ProposalProgressBar';
+import ProposalProgressBar from 'containers/InitiativesShow/ReactionControl/Status/components/ProposalProgressBar';
 
 import { isNilOrError } from 'utils/helperUtils';
