--- front/app/containers/InitiativesShow/ReactionControl/ProposalProgressBar.tsx
@@ -1,54 +1,0 @@
-import React from 'react';
-import styled, { useTheme } from 'styled-components';
-import ProgressBar from 'components/UI/ProgressBar';
-import { ScreenReaderOnly } from 'utils/a11y';
-import { FormattedMessage } from 'utils/cl-intl';
-import messages from './messages';
-import { colors } from '@citizenlab/cl2-component-library';
-
-const StyledProgressBar = styled(ProgressBar)`
-  height: 12px;
-  width: 100%;
-`;
-
-interface Props {
-  reactionCount: number;
-  reactionLimit: number;
-  barColor?: string;
-  className?: string;
-}
-
-const ProposalProgressBar = ({
-  reactionCount,
-  reactionLimit,
-  barColor,
-  className,
-}: Props) => {
-  const theme = useTheme();
-
-  return (
-    <div className={className}>
-      <StyledProgressBar
-        progress={reactionCount / reactionLimit}
-        color={barColor || theme.colors.tenantText}
-        bgColor={colors.grey200}
-      />
-      <ScreenReaderOnly>
-        <FormattedMessage
-          {...messages.a11y_xVotesOfRequiredY}
-          values={{
-            xVotes: (
-              <FormattedMessage
-                {...messages.xVotes}
-                values={{ count: reactionCount }}
-              />
-            ),
-            votingThreshold: reactionLimit,
-          }}
-        />
-      </ScreenReaderOnly>
-    </div>
-  );
-};
-
-export default ProposalProgressBar;
