+++ front/app/containers/InitiativesShow/ReactionControl/Status/components/ReactionCounter.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import styled from 'styled-components';
 import messages from '../../messages';
 import { FormattedMessage } from 'utils/cl-intl';
-import ProposalProgressBar from '../../ProposalProgressBar';
+import ProposalProgressBar from './ProposalProgressBar';
 import { StatusComponentProps } from '../../StatusWrapper';
 
 const ReactionText = styled.div`
